Added ability to drag/drop new entries in tabular inline sets.
In order to implement this, I needed to make two changes, one of
which is fairly significant.

1) The javascript no longer limits the drag mechanism to <td>'s with
the "has_original" class.
2) The CustomInlineFormSet.save_new() function now checks default_sort_order
to see whether it's 0, intead of checking its type. This necessitated a change
to the documentation, stating that "default=0" must be added to the sort field.

The default has to be zero for two reasons:

1) It simplifies that check in CustomInlineFormSet.save_new(). A default of 0 is
compatible with the javascript, since it uses 1-indexed sorting.
2) Users of the newest South versions are forced to provide a default value when
they add a new sort field to their model.

Ultimately, this change allows us to know for certain what the default value will
be for a new inline entry which has not yet been dragged, allowing the max_order
algorithm to work properly in all cases.