Fixes #162
This package appears to depend on their existing admin actions, which, in turn trigger the rendering of per-item checkboxes on the change-list.

The dependency is simply a means of obtaining the item ID, so that the frontend can be updated with JS after a move. It was trivial to instead, render a custom class on our custom sort widget that provides the same ID, thus removing a dependency for an unrelated feature.