Update admin.py
When an inline does not have any fields listed, everything works fine. But when fields does exist and does not include the order field, the order field is not added at all, so no changes ever happen to the ordering when saved. This fix should remedy that, although it's untested. Either that, or a note should be made to always include the order field in the field list, which is what I will be doing until this is fixed.