Use order_field’s position for sort field, if available, but don’t include both
It is sometimes desirable to be able to specify the column position of the resulting ordering widgets in the ModelAdmin’s change list by manually including the order field in `list_display`. However, it is not desireable to display both the widget and the value of the order field. This PR addresses this.